third_party/blink/web_tests/platform/generic/http/tests/security/cross-origin-createImageBitmap-expected.txt

The image bitmap factories should not throw exceptions on cross-origin access.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Resolved as expected: image
PASS ImageBitmap is tainted. Threw error: SecurityError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The canvas has been tainted by cross-origin data.
PASS Resolved as expected: canvas
PASS ImageBitmap is tainted. Threw error: SecurityError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The canvas has been tainted by cross-origin data.
PASS Resolved as expected: imageBitmap
PASS ImageBitmap is tainted. Threw error: SecurityError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The canvas has been tainted by cross-origin data.
PASS Resolved as expected: video
PASS ImageBitmap is tainted. Threw error: SecurityError: Failed to execute 'getImageData' on 'CanvasRenderingContext2D': The canvas has been tainted by cross-origin data.
PASS successfullyParsed is true

TEST COMPLETE
